refactor: move json_merge_patch out of config/utils into its own module
